update v6 - Added the re-assignment ability when deparment/jobs is deleted. Updated Models, Seeds, Routes

+++ controllers/api/jobs-Routes.js
@@ -1,12 +1,20 @@
 const router = require('express').Router();
-const { Jobs } = require('../../models');
+const { Department, Jobs, Employee } = require('../../models');
 
 //===========================================================================
 
 // GET all jobs
 router.get('/', async (req, res) => {
     try {
-        const jobsData = await Jobs.findAll();
+        const jobsData = await Jobs.findAll({
+            include: [
+                {
+                    model: Department,
+                    attributes: ['id', 'department_name']
+                }
+            ]
+        });
+
         res.status(200).json(jobsData); 
     } catch (err) {
         res.status(500).json(err);
@@ -18,7 +26,15 @@ router.get('/', async (req, res) => {
 // GET a single job by id
 router.get('/:id', async (req, res) => {
     try {
-        const jobData = await Jobs.findByPk(req.params.id);
+        const jobData = await Jobs.findByPk(req.params.id, {
+            include: [
+                {
+                    model: Department,
+                    attributes: ['id', 'department_name']
+                }
+            ]
+        });
+
         res.status(200).json(jobData);
     } catch (err) {
         res.status(500).json(err);
@@ -64,12 +80,26 @@ router.put('/:id', async (req, res) => {
 router.delete('/:id', async (req, res) => {
     try {
         const jobsToDelete = await Jobs.findByPk(req.params.id);    // Fetch the `Jobs` model. Stores data in the `jobsToDelete` TEMPORARILY for only this function router.delete(). Once function ends, `jobsToDelete` will be garbaged.
-
         if (!jobsToDelete) {
             res.status(404).json({ message: 'No job found with this id!' });
             return;
         }
 
+        // Fetch the `Unassigned Department`
+        const unassignedJobs = await Jobs.findOne({
+            where: { job_title: 'Unassigned Jobs (DO NOT REMOVE)'}
+        });
+        if (!unassignedJobs) {
+            res.status(404).json({ message: 'No Unassigned Jobs found!' });
+            return;
+        };
+
+        // Find All Employees to Job ID
+        await Employee.update({ job_id: unassignedJobs.id }, {
+            where: { job_id: req.params.id },
+        });
+
+        // Delete Jobs
         await Jobs.destroy({
             where: { id: req.params.id },
         });
